exercise 4.21:  change delete - only creator can delete

+++ part4/bloglist/controllers/blogs.js
@@ -45,6 +45,18 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blogToDelete = await Blog.findById(request.params.id)
+
+  if (blogToDelete.user.toString() !== decodedToken.id.toString()) {
+    return response.status(401).json({ error: 'unauthorized to delete ' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
